+++ app/Http/Controllers/ProductController.php
@@ -2,10 +2,9 @@
 
 namespace App\Http\Controllers;
 
-use App\Product;
-use App\User;
-use Illuminate\Http\Request;
+use App\Models\Product;
 use App\Services\ProductService;
+use Illuminate\Http\Request;
 
 class ProductController extends Controller
 {
