+++ app/Services/ProductService.php
@@ -2,7 +2,7 @@
 
 namespace App\Services;
 
-use App\Product;
+use App\Models\Product;
 
 class ProductService
 {
